Refactor Mask in Attention (#1028)
Currently, the attention code is **hardcoded** to handle either `causal_mask`
or an arbitrary `mask_fn`.

To support **sliding window masks**, we previously used a **hack** by injecting
the `_sliding_window_size` attribute into functions.

This refactor **makes the masking logic more flexible** by allowing arbitrary
`MaskFnAttentionBias`.
- If downstream requires a **new mask pattern**, they can simply:
  1. Implement a **subclass of `MaskFnAttentionBias`**.
  2. Set `attention.mask` accordingly.

+++ axlearn/common/multiway_transformer_test.py
@@ -139,7 +139,7 @@ class ModelTest(parameterized.TestCase):
             prng_key=jax.random.PRNGKey(0),
         )
         initial_state, initial_output = layer.init_states(
-            time_step=None, data=TensorSpec([batch_size, tgt_len])
+            time_step=None, data=TensorSpec([batch_size, tgt_len], dtype=jnp.float32)
         )
         self.assertIsNone(initial_output)
         inputs = dict(
